changes console output, class distribution of methods and minor fixes
